+++ bugsnag-plugin-android-ndk/src/main/jni/utils/stack_unwinder.cpp
@@ -1,5 +1,6 @@
 #include "stack_unwinder.h"
 
+#include "logger.h"
 #include "string.h"
 
 #include <dlfcn.h>
@@ -53,7 +54,7 @@ void bsg_unwinder_init() {
   }
   attempted_init = true;
 
-  auto crash_time_maps = new unwindstack::LocalMaps();
+  auto crash_time_maps = new unwindstack::LocalUpdatableMaps();
   if (crash_time_maps->Parse()) {
     std::shared_ptr<unwindstack::Memory> crash_time_memory(
         new unwindstack::MemoryLocal);
@@ -69,6 +70,58 @@ void bsg_unwinder_init() {
   if (!current_time_unwinder->Init()) {
     delete current_time_unwinder;
     current_time_unwinder = nullptr;
+  }
+}
+
+static void populate_code_identifier(const unwindstack::FrameData &frame,
+                                     bugsnag_stackframe &dst_frame) {
+  auto maps = crash_time_unwinder->GetMaps();
+  if (maps == nullptr) {
+    return;
+  }
+
+  auto map_info = maps->Find(frame.pc);
+  if (map_info == nullptr) {
+    return;
+  }
+
+  auto elf_fields = map_info->elf_fields();
+  if (elf_fields == nullptr) {
+    return;
+  }
+
+  auto shared_build_id = elf_fields->build_id_.load();
+  std::string_view build_id;
+  if (shared_build_id == nullptr) {
+    auto elf = elf_fields->elf_.get();
+    if (elf == nullptr) {
+      return;
+    }
+
+    build_id = elf->GetBuildID();
+  } else {
+    build_id = *shared_build_id;
+  }
+
+  if (build_id.empty()) {
+    return;
+  }
+
+  // MapInfo.GetPrintableBuildID is *not* async-safe so we need our own
+  // safe hex encoder to copy BuildID into code_identifier.
+  bsg_hex_encode(dst_frame.code_identifier, build_id.data(), build_id.length(),
+                 sizeof(dst_frame.code_identifier));
+}
+
+void bsg_unwinder_refresh(void) {
+  if (crash_time_unwinder == nullptr) {
+    return;
+  }
+
+  auto *crash_time_maps = dynamic_cast<unwindstack::LocalUpdatableMaps *>(
+      crash_time_unwinder->GetMaps());
+  if (crash_time_maps != nullptr) {
+    crash_time_maps->Reparse(nullptr);
   }
 }
 
@@ -95,6 +148,8 @@ ssize_t bsg_unwind_crash_stack(bugsnag_stackframe stack[BUGSNAG_FRAMES_MAX],
     dst_frame.line_number = frame.rel_pc;
     dst_frame.load_address = frame.map_start;
     dst_frame.symbol_address = frame.pc - frame.function_offset;
+
+    populate_code_identifier(frame, dst_frame);
 
     // if the filename or method name cannot be found (or are considered
     // invalid) - fallback to dladdr to find them
@@ -131,6 +186,10 @@ bsg_unwind_concurrent_stack(bugsnag_stackframe stack[BUGSNAG_FRAMES_MAX],
       dst_frame.line_number = frame.rel_pc;
       dst_frame.load_address = frame.map_info->start();
       dst_frame.symbol_address = frame.pc - frame.map_info->offset();
+
+      bsg_strncpy(dst_frame.code_identifier,
+                  frame.map_info->GetPrintableBuildID().c_str(),
+                  sizeof(dst_frame.code_identifier));
 
       // if the filename is empty or invalid, use the `Elf` info to get the
       // correct Soname if the function_name is empty as well, fallback to