+++ features/fixtures/mazerunner/cxx-scenarios-bugsnag/src/main/cpp/CXXExternalStackElementScenario.cpp
@@ -1,6 +1,7 @@
 #include <jni.h>
 #include <cstdio>
 #include <dlfcn.h>
+#include <bugsnag.h>
 
 extern "C" {
 // defined in libs/[ABI]/libmonochrome.so
@@ -10,6 +11,7 @@ JNIEXPORT int JNICALL
 Java_com_bugsnag_android_mazerunner_scenarios_CXXExternalStackElementScenario_crash(
     JNIEnv *env, jobject instance, jint counter) {
   void *monochrome = dlopen("libmonochrome.so", RTLD_GLOBAL);
+  bugsnag_refresh_symbol_table();
   *(void**)(&something_innocuous) = dlsym(monochrome, "something_innocuous");
 
   printf("Captain, why are we out here chasing comets?\n%d\n", counter);